Add Appstream metainfo file
This is a starting point for an Appstream metainfo file. It probably needs to
go via templating to set the installation paths. It also need to get installed
by cmake: /usr/share/metainfo/projecteur.metainfo.xml is the usual location.
The Id field and filename should normally be {tld}.{vendor}.{product} and that
would also be reflected in the name of the .desktop file IIRC.

https://www.freedesktop.org/software/appstream/docs/chap-Metadata.html