Fix InputMapper issue #144
In case, when two different input sequence of same length and same end
KeyEvent are defined then InputMapper fails to map to correct action.

This commit fixes the issue with InputMapper. The correct action is
decided on the basis of all keys pressed (instead of last key pressed
in earlier version of code).