Add button release code for button long press events
This commit add the button release code for signifying the end of
long-press event. While presenting, if a user press next/back button for
a while and then release it within `Input Sequence Interval` time (a
setting in Input Mapper tab), only the action corresponding to
long-press will be executed.

Hence, now there are three specific actions possible for each of Next
and Hold button on Logitech Spotlight:
1. Tap event
2. Long-Press (but not longer than `Input Sequence Interval` setting in Input Mapper tab)
3. Hold Move: It can be mapped to special motion related mapped actions
   like `Scrolling` (Horizontal and Vertical) or `Volume Control`.